refactor: split dispatcher
